Separate server code into modules. Add dotenv.
