--- socket/src/model.d.ts
@@ -1,9 +1,0 @@
-export type Player = {
-    socketId: string
-    roomId: string
-    name?: string,
-}
-
-export type Room = {
-    roomId: string
-}
