+++ socket/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
-        "target": "es6",
-        "module": "es2015",
+        "target": "es2022",
+        "module": "es2022",
         "moduleResolution": "node",
         "strict": true,
         "esModuleInterop": true,
